Move interop test client to its own package
