+++ Dockerfile.interop_client
@@ -16,5 +16,8 @@ COPY lerna.json .
 COPY nx.json .
 COPY packages/ packages/
 COPY scripts/ scripts/
+# Re-run clean-install to set up bin symlinks.
+RUN npm ci
+RUN npx lerna run build --scope interop-test-client --no-progress && rm -rf node_modules/.cache/nx
 
-CMD ["/bin/sh", "-c", "npm exec -- ts-node scripts/interopTestClient.ts >/logs/stdout.log 2>/logs/stderr.log"]
+CMD ["/bin/sh", "-c", "npx interop-test-client >/logs/stdout.log 2>/logs/stderr.log"]
