+++ packages/interop-test-client/bin/interop-test-client
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+require("../dist/index.js").interopTestClient();
