+++ packages/interop-test-client/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "interop-test-client",
+  "version": "0.0.0",
+  "private": true,
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "source": "src/index.ts",
+  "module": "dist/module.js",
+  "bin": {
+    "interop-test-client": "bin/interop-test-client"
+  },
+  "files": [
+    "bin",
+    "src"
+  ],
+  "scripts": {
+    "clean": "rm -rf dist/*",
+    "build": "parcel build"
+  },
+  "dependencies": {
+    "@divviup/dap": "^0.1.0",
+    "express": "^4.18.1"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.14"
+  }
+}
